+++ src/scripts.js
@@ -7,6 +7,8 @@ import './images/Sleep.svg';
 import './images/Friends.svg';
 import './images/activity.svg'
 import { getUserData, getSleepData, getActivityData, getHydrationData } from './fetch';
+import { generateFlightsChart, generateActivityChart, generateWaterChart, generateSleepChart } from './charts';
+// import { } from './domManipulation';
 import DataManager from './DataManager';
 import UserRepository from './UserRepository';
 import { Chart, registerables } from 'chart.js';
@@ -15,13 +17,10 @@ Chart.register(...registerables);
 // Query Selectors
 const userProfile = document.querySelector('#userProfile');
 const todaySteps = document.querySelector('#stepsToday');
-// const activityGoals = document.querySelector('#activityGoals');
 const stepGoals = document.querySelector('#stepGoals');
 const waterStats = document.querySelector('#waterStats');
-const waterCalendar = document.querySelector('#waterCanvas').getContext('2d');
 const sleepHours = document.querySelector('#sleepHours');
 const sleepQuality = document.querySelector('#sleepQuality');
-const sleepCalendar = document.querySelector('#sleepCanvas').getContext('2d');
 const friendContainer = document.querySelector('#FriendDisplay');
 // const scrollSleep = document.getElementById('scrollSleep')
 
@@ -117,6 +116,8 @@ const displayHydrationInfo = (user) => {
   <p>oz</p>
   `;
   generateWaterChart(user);
+  generateActivityChart(user)
+  generateFlightsChart(user)
 };
 
 const displaySleepInfo = (user) => {
@@ -145,113 +146,5 @@ const displayFriendsInfo = (user) => {
     friendContainer.innerHTML += `<p>${friend.name}</p>`
   });
 }
-
-const generateWaterChart = (user) => {
-  const waterChart = new Chart(waterCalendar, {
-    type: 'line',
-    data: {
-      labels: ['M', 'T', 'W', 'Th', 'Fr', 'Sa', 'Su'],
-      datasets: [{
-        label: 'oz of water',
-        data: user.hydrationData.getWeeklyDrank('2020/01/15'),
-        backgroundColor: '#FC6F7F',
-        borderColor: '#FC6F7F',
-        borderWidth: 2
-      }]
-    },
-    options: {
-      plugins: {
-        legend: {
-          display: true,
-          position: 'bottom',
-          labels: {
-            color: '#ffffff'
-          },
-        }
-      },
-      scales: {
-        y: {
-          ticks: {
-            color: ['#ffffff'],
-          },
-          beginAtZero: true,
-          grid: {
-            color: 'rgba(0, 0, 0, 0)',
-            borderColor: 'rgba(0, 0, 0, 0)',
-          }
-        },
-        x: {
-          ticks: {
-            color: ['#ffffff']
-          },
-          grid: {
-            color: 'rgba(0, 0, 0, 0)',
-            borderColor: 'rgba(0, 0, 0, 0)',
-          },
-        },
-      }
-    },
-  })
-};
-
-const generateSleepChart = (user) => {
-  const sleepChart = new Chart(sleepCalendar, {
-    type: 'line',
-    data: {
-      labels: ['M', 'T', 'W', 'Th', 'Fr', 'Sa', 'Su'],
-      datasets: [
-        {
-          label: 'hours',
-          data: user.sleepData.getWeeklyHoursSlept('2020/01/15'),
-          backgroundColor: '#FC6F7F',
-          borderColor: '#FC6F7F',
-          borderWidth: 2
-        },
-        {
-          label: 'quality',
-          data: user.sleepData.getWeeklySleepQuality('2020/01/15'),
-          backgroundColor: '#FF9E2D',
-          borderColor: '#FF9E2D',
-          borderWidth: 2
-        }]
-    },
-    options: {
-      plugins: {
-        legend: {
-          display: true,
-          position: 'bottom',
-          labels: {
-            color: '#ffffff',
-            padding: 15,
-          },
-        }
-      },
-      scales: {
-        y: {
-          ticks: {
-            color: ['#ffffff'],
-            stepSize: 6,
-          },
-          beginAtZero: true,
-          min: 0,
-          max: 12,
-          grid: {
-            color: 'rgba(0, 0, 0, 0)',
-            borderColor: 'rgba(0, 0, 0, 0)',
-          }
-        },
-        x: {
-          ticks: {
-            color: ['#ffffff'],
-          },
-          grid: {
-            color: 'rgba(0, 0, 0, 0)',
-            borderColor: 'rgba(0, 0, 0, 0)',
-          },
-        },
-      }
-    },
-  })
-};
 
 retrieveAllData();
